chore: refactor subquery
